:dizzy: /api/settings → routes/api/settings.js
